SystemSettings: load 'custom' rdsampjni if possible.
It's not possible (at least, it's not easy) to access the
WFDB_CUSTOMLIB parameter at the time rdsampjni is loaded.

So, instead, simply try to load the custom one, and if that fails, try
to load the prebuilt one.  If a custom rdsampjni exists and is
loadable, safe to assume that it should take precedence over the
prebuilt one.