Assortment of build system improvements and a few bug fixes. (#141)
* Remove various generated files.

* wrann: remove fake date/time from text input.

Do not use '00:00.000' as a fake time stamp; that is likely to cause
confusion and breakage in the future.  Use '---------' which makes it
obvious that this is not a real time stamp, and that its length (9
characters) is necessary to accomodate wrann historical weirdness.

* rdmat: avoid use of 'containers.Map'.

* Remove assorted bits of junk that do not belong here.

All relevant issues have been resolved in the upstream packages.

* UnitTests: use new rdsamp API.

* wfdbloadlib: Use the same jar file regardless of the Java version.

For official releases, we will build *one* jar file, using the oldest
jdk version that we support.  I don't know of any reason that this jar
file will not work with newer jvm versions.

* rdsampjni: Fix headers.

Remove nonstandard <malloc.h> which is not used.

Include generated "org_physionet_wfdb_jni_Rdsamp.h" to ensure
compatibility with the expected prototype.

* rdsampjni: new autoconf+libtool build system.

* SystemSettings: use standard paths for loading rdsampjni.

* Makefile for building nativelibs tools, with many improvements.

Instead of multiple ad-hoc makefiles, unify everything into one file.

All external sources must be verified.

Add additional rules for cross-compiling (note, however, that these
require various standard and non-standard cross build tools to be
configured separately.)

Note that wfdb-10.5.25pre2 is currently broken with this makefile.

* Makefile: assorted cleanup.

* More stuff to ignore.

* SystemSettings: load 'custom' rdsampjni if possible.

It's not possible (at least, it's not easy) to access the
WFDB_CUSTOMLIB parameter at the time rdsampjni is loaded.

So, instead, simply try to load the custom one, and if that fails, try
to load the prebuilt one.  If a custom rdsampjni exists and is
loadable, safe to assume that it should take precedence over the
prebuilt one.

* zipexclude.lst: More stuff to exclude from the zip file.

* Makefile: automatically update WFDB_CUSTOMLIB.

* Makefile: add 'make check' target.

* nativelibs/Makefile: remove extraneous prebuilt binaries.

Also fix the issue of trying to lipo two .la files.

* zipexclude.lst: correct syntax for excluding subdirectories.

* Rdsamp: Do not try to load liblibwfdb.

This line must have been added at some point in an attempt to fix a
broken version of librdsampjni that didn't import its dependencies
correctly.

On all supported platforms, shared libraries are responsible for
managing their own dependencies.