Wfdbexec: handle standard error output.
If we launch a child process using ProcessBuilder with
redirectErrorStream set to false, the process's standard output and
standard error will be connected to two separate pipes.

This is generally a good thing, as it allows us to parse the program's
output without trying to parse error messages.  However, it also means
that we must actually read from both pipes (since pipe buffers are
finite, if the child process writes enough text to a pipe without
anyone reading from the other end, the process will block.)  We cannot
predict in what order the child process will write output and error
messages, or how large those messages will be, so we need to be able
to read from the two pipes in any order.  Possibly the simplest way to
do so is to launch a separate thread to read from one of the pipes
(standard error), while continuing to read from the other pipe
(standard output) in the main thread.

Moreover, in many cases the error messages printed by WFDB
applications are actually meaningful and should not simply be ignored.

With this change, error messages from the child process will be read
as they are written, and sent to the logger at level WARNING.
(Currently, this means that the messages will be hidden unless
wfdbloadlib is invoked with debug level 2 or higher.)

The following functions will now send standard error messages to the
logger:

   execToStringList()
   execToStringList(String[])
   execToDoubleArray(String[])
   execToDoubleList(String[])
   execToFloatArray(String[])
   execToLongArray(String[])
   execToShortArray(String[])

The following functions set redirectErrorStream to true, and thus
standard error and standard output are combined into the single list
returned to the caller:

   execWithStandardInput(String[])
   execWithStandardInput(byte[])