bxb: explicitly define the report file parsing.
'importdata' is ... underspecified, to put it mildly, and at least in
octave 4.0.3, it doesn't do anything useful with bxb report files.

The expected output format for the bxb function isn't documented, but
I've tried to reconstruct what the output ought to look like, and
written a parser that does just that.

I have not bothered trying to reconstruct 'textdata' because it's hard
to see how that could ever be useful to anyone.  If somebody actually
wants 'textdata', feel free to write a spec for what it ought to
contain!