wfdbloadlib: use normal stack size for process reaper threads.
This is to fix cases where Wfdbexec either:

(a) fails with an OutOfMemoryError, or

(b) displays a warning about a StackOverflowError, and then
    (sometimes, depending on which Wfdbexec method is used) causes the
    entire process to hang.

In current versions of glibc, the "thread stack" memory contains both
the actual stack, and the static thread-local storage area.  The size
that is passed to pthread_attr_setstacksize must include space for
both (even though there's no portable way for a library to determine
how large the process's thread-local storage is.)  See
<https://sourceware.org/bugzilla/show_bug.cgi?id=11787>.

This is a problem because, when a Java program creates a child
process, the JVM also creates an internal thread to monitor the
process's state, and since that thread doesn't do any actual
processing, it is set to use a small stack size by default (currently
128 kilobytes plus some overhead.)  The JVM has no idea how large the
process's thread-local storage is, and simply assumes that 128k is
enough.  This is not the case, at least, in current versions of Octave
on Debian or Ubuntu.

To work around this, set the system property
jdk.lang.processReaperUseDefaultStackSize to true.  This
confusingly-named property causes the JVM to use the "default thread
stack size" (typically 1 MB) when creating a process reaper thread,
instead of the process-reaper default of 128k.

This will consume slightly more virtual memory, but given that we are
only trying to run one child process at a time (and we no longer have
official support for 32-bit architectures) this is not a significant
concern.

Future versions of Octave will do the same thing when they detect they
are running on a version of glibc that is susceptible (see
https://savannah.gnu.org/bugs/?58641).