wfdbjava2mat: new function.
This function takes an array of numeric values (1D or 2D) that is
returned by a Java function, and converts it to a matrix.

This conversion is automatic in Matlab, so this function is a no-op
there.  In Octave, historically, this was done using the 'java2mat'
function, although in newer versions of Octave, automatic conversion
is performed for 1D (but not 2D) arrays.  Newer versions of Octave
have renamed 'java2mat' to '__java2mat__' and deprecated the older
name.