Don't cache the result of getWfdbClass.
Previously, each function that used getWfdbClass would cache the
result; this meant the object did not have to be re-created each time,
but it also meant that if the configuration were changed using
wfdbloadlib, the old configuration (particularly DEBUG_LEVEL and
NETWORK_WAIT_TIME) would continue to be used.

We want to be able to change the configuration at runtime, so don't
cache the result of getWfdbClass.  Instead, call getWfdbClass again
each time it is needed.